refactor: Implement useRequestPasswordReset hook
